Force warp matrix to be float32 data type

+++ opendm/multispectral.py
@@ -511,7 +511,7 @@ def find_ecc_homography(image_gray, align_image_gray, number_of_iterations=1000,
     log.ODM_INFO("Pyramid levels: %s" % pyramid_levels)
 
     if warp_matrix_init is not None:
-        warp_matrix = warp_matrix_init * np.array([[1,1,2],[1,1,2],[0.5,0.5,1]], dtype=np.float32)**(1-(pyramid_levels+1))
+        warp_matrix = warp_matrix_init.astype(np.float32) * np.array([[1,1,2],[1,1,2],[0.5,0.5,1]], dtype=np.float32)**(1-(pyramid_levels+1))
     else:
         # Quick check on size
         if align_image_gray.shape[0] != image_gray.shape[0]:
@@ -563,7 +563,7 @@ def find_ecc_homography(image_gray, align_image_gray, number_of_iterations=1000,
             if level != pyramid_levels:
                 log.ODM_INFO("Could not compute ECC warp_matrix at pyramid level %s, resetting matrix" % level)
                 if warp_matrix_init is not None:
-                    warp_matrix = warp_matrix_init * np.array([[1,1,2],[1,1,2],[0.5,0.5,1]], dtype=np.float32)**(1-(pyramid_levels+1))
+                    warp_matrix = warp_matrix_init.astype(np.float32) * np.array([[1,1,2],[1,1,2],[0.5,0.5,1]], dtype=np.float32)**(1-(pyramid_levels+1))
                 else:
                     warp_matrix = np.eye(3, 3, dtype=np.float32)
                     warp_matrix = warp_matrix * np.array([[1,1,2],[1,1,2],[0.5,0.5,1]], dtype=np.float32)**(1-(pyramid_levels+1))
